refactor(apps/web): migrate components to components dir
