+++ apps/web/src/components/Footer/index.tsx
