+++ apps/web/src/main.tsx
@@ -5,8 +5,8 @@ import { createRoot } from "react-dom/client";
 import { Provider } from "react-redux";
 import { BrowserRouter as Router } from "react-router-dom";
 
-import App from "./components/App";
 import "./index.css";
+import App from "./layout";
 import { AuthProvider } from "./lib/auth";
 import { store } from "./lib/store";
