+++ apps/web/src/scenes/_authenticated.tsx
@@ -2,8 +2,8 @@ import { useApiIsLoaded } from "@vis.gl/react-google-maps";
 import { useEffect } from "react";
 import { Route, Routes } from "react-router-dom";
 
+import Toast from "src/components/Toast";
 import { useAuth } from "src/lib/auth";
-import { useAppDispatch } from "src/lib/store";
 import * as ROUTES from "src/lib/utils/routes";
 import Account from "src/scenes/account";
 import Dashboard from "src/scenes/dashboard";
@@ -13,8 +13,6 @@ import ListInvites from "src/scenes/list/invites";
 import Lists from "src/scenes/lists";
 import NotFound from "src/scenes/not-found";
 import Place from "src/scenes/place";
-
-import Toast from "../../Toast";
 
 const AuthenticatedRoutes = () => {
 	const { setCurrentLocation } = useAuth();
