+++ apps/web/src/scenes/bookmarks/index.tsx
@@ -1,11 +1,10 @@
 import Loading from "@hominem/components/Loading";
 import { useNavigate } from "react-router-dom";
 
+import BookmarkForm from "src/components/BookmarkForm";
+import BookmarkListItem from "src/components/BookmarkListItem";
 import { useGetBookmarks } from "src/lib/api/bookmarks";
 import { useAuth } from "src/lib/auth";
-
-import BookmarkForm from "./components/BookmarkForm";
-import BookmarkListItem from "./components/BookmarkListItem";
 
 const Recommendations = () => {
 	const navigate = useNavigate();
