+++ apps/web/src/scenes/dashboard/index.tsx
@@ -1,7 +1,8 @@
 import React, { Suspense } from "react";
 
 const LazyDashboard = React.lazy(
-	() => import(/* webpackChunkName: "dashboard" */ "./Dashboard"),
+	() =>
+		import(/* webpackChunkName: "dashboard" */ "../../components/Dashboard"),
 );
 
 const Dashboard = (props) => (
