+++ apps/web/src/scenes/list/invites/index.tsx
@@ -6,7 +6,7 @@ import { Link, useNavigate, useParams } from "react-router-dom";
 import { useGetList, useGetListInvites } from "src/lib/api";
 import { useAuth } from "src/lib/auth";
 
-import ListInviteForm from "./components/ListInviteForm";
+import ListInviteForm from "../../../components/ListInviteForm";
 
 const ListInvites = () => {
 	const params = useParams();
