+++ apps/web/src/scenes/lists/index.tsx
@@ -6,8 +6,8 @@ import { useNavigate } from "react-router-dom";
 import { useGetLists } from "src/lib/api";
 import { useAuth } from "src/lib/auth";
 
-import ListForm from "./components/ListForm";
-import Lists from "./components/Lists";
+import ListForm from "../../components/ListForm";
+import Lists from "../../components/Lists";
 
 const ListsRoute = () => {
 	const navigate = useNavigate();
